Add fonts for international scripts
This also adds the Preview feature on the ReportBro designer.

The main fonts that were added:
* GNU Unifont: supports most non latin characters but is a bitmap font intended for screen display
* GNU FreeFont: Versatile font with good support for Devanagari (Nepalese)
* Firefly: Generic font (mostly latin)
* Google Noto Sans: Open Source Font Family with support for hundreds of scripts. Each variant needs to be installed separately. Included for now are Arabic (regular and Naskh), Myanmar, Ethiopic and general Noto Sans.

Some important caveats: ReportBro is based on the Pillow library that has poor support for right-to-left languages. Arabic will therefore have to be reversed with python-bidi or other tool and will still have some ligature issues. Myanmar seems to introduce line breaks on accents.
All contributions are welcome to improve this.